# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Veterinary Management
  description: Veterinary Management
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: "Local server"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication using OAuth2 Bearer token.

        **How to authenticate:**
        1. POST to `/loginJWT` with `administrator_id` and `password`
        2. Receive JWT token in response
        3. Include token in Authorization header: `Bearer <token>`

        **Available scopes:**
        - `RECEPTIONIST` - Clinic receptionist staff (manage visits, view schedules)
        - `CLINIC_MANAGER` - Clinic manager (full system access)
        - `VETERINARIAN` - Veterinarian (manage own schedule, view assigned visits)
        - `PET_OWNER` - Pet owner (view own pets and visits)
  parameters:
    vet_id:
      name: vet_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/VeterinarianId'
    availability_id:
      name: availability_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/AvailabilityId'
    exception_id:
      name: exception_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ExceptionId'
...
    start_date:
      name: start
      in: query
      required: true
      description: "Start date of the requested schedule (inclusive)"
      schema:
        type: string
        format: date
      default: "2025-09-01"
    end_date:
      name: end
      in: query
      required: true
      description: "End date of the requested schedule (inclusive)"
      schema:
        type: string
        format: date
      default: "2025-10-01"
  requestBodies:
    AvailabilityInput:
      description: "Availability object input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              day_of_week:
                type: integer
              start_time:
                $ref: '#/components/schemas/LocalTime'
              end_time:
                $ref: '#/components/schemas/LocalTime'
              period_start:
                $ref: '#/components/schemas/LocalDate'
              period_end:
                $ref: '#/components/schemas/LocalDate'
    AvailabilityExceptionInput:
      description: "Availability exception input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              reason:
                type: string
              day_of_week:
                type: integer
              start_time:
                $ref: '#/components/schemas/LocalTime'
              end_time:
                $ref: '#/components/schemas/LocalTime'
              period_start:
                $ref: '#/components/schemas/LocalDate'
              period_end:
                $ref: '#/components/schemas/LocalDate'
...
  schemas:
    LocalTime:
      type: string
      format: time
      description: "Local time in the format HH:MM"
      default: "09:00"
    LocalDate:
      type: string
      format: date
      description: "Local date in the format YYYY-MM-DD"
      default: "2025-01-01"
    VeterinarianId:
      type: integer
      format: int64
      description: "Unique identifier for a Veterinarian"
    AvailabilityId:
      type: integer
      format: int64
      description: "Unique identifier for an Availability"
    ExceptionId:
      type: integer
      format: int64
      description: "Unique identifier for an Exception"
...
    MedicationId:
      type: integer
      format: int64
      description: "Unique identifier for an Medication"
...
    VisitId:
      type: integer
      format: int64
      description: "Unique identifier for a Visit"
    PrescriptionId:
      type: integer
      format: int64
      description: "Unique identifier for a Medication Prescription"
    PetId:
      type: integer
      format: int64
      description: "Unique identifier for a Pet"
    PetOwnerId:
      type: integer
      format: int64
      description: "Unique identifier for a Pet Owner"
    TreatmentId:
      type: integer
      format: int64
      description: "Unique identifier for a Treatment"
    Veterinarian:
      type: object
      properties:
        vet_id:
          $ref: '#/components/schemas/VeterinarianId'
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        email:
          type: string
        address:
          type: string
        license_number:
          type: string
        years_of_experience:
          type: integer
        availabilities:
          type: array
          items:
            $ref: '#/components/schemas/Availability'
    Availability:
      type: object
      properties:
        availability_id:
          $ref: '#/components/schemas/AvailabilityId'
        day_of_week:
          type: integer
        start_time:
          $ref: '#/components/schemas/LocalTime'
        end_time:
          $ref: '#/components/schemas/LocalTime'
        period_start:
          $ref: '#/components/schemas/LocalDate'
        period_end:
          $ref: '#/components/schemas/LocalDate'
        veterinarian:
          $ref: '#/components/schemas/Veterinarian'
    AvailabilityException:
      type: object
      properties:
        exception_id:
          $ref: '#/components/schemas/ExceptionId'
        reason:
          type: string
        day_of_week:
          type: integer
        start_time:
          $ref: '#/components/schemas/LocalTime'
        end_time:
          $ref: '#/components/schemas/LocalTime'
        period_start:
          $ref: '#/components/schemas/LocalDate'
        period_end:
          $ref: '#/components/schemas/LocalDate'
        availability:
          $ref: '#/components/schemas/Availability'
...
    Visit:
      type: object
      properties:
        visit_id:
          $ref: '#/components/schemas/VisitId'
        visit_date:
          $ref: '#/components/schemas/LocalDate'
        visit_time:
          $ref: '#/components/schemas/LocalTime'
        duration:
          type: integer
        reason_for_visit:
          type: string
        diagnoses:
          type: string
        notes:
          type: string
        price_per_fifteen:
          type: double
        status:
          $ref: '#/components/schemas/VisitStatus'
        veterinarian_id:
          $ref: '#/components/schemas/VeterinarianId'
        pet_id:
          $ref: '#/components/schemas/PetId'
        pet_owner_id:
          $ref: '#/components/schemas/PetOwnerId'
        medication_prescriptions:
          type: array
          items:
            $ref: '#/components/schemas/MedicationPrescriptionInformation'
        treatment:
          $ref: '#/components/schemas/Treatment'
...
    MedicationPrescriptionInformation:
      type: object
      properties:
        prescription_id:
          $ref: '#/components/schemas/PrescriptionId'
        visit_id:
          $ref: '#/components/schemas/VisitId'
        medication_id:
          $ref: '#/components/schemas/MedicationId'
        medication_name:
          type: string
        quantity:
          type: integer
        dosage_instructions:
          type: string
        duration_in_days:
          type: integer
    Treatment:
      type: object
      properties:
        treatment_id:
          $ref: '#/components/schemas/TreatmentId'
        name:
          type: string
        description:
          type: string
        cost:
          type: number
          format: double
...
    VisitStatus:
      type: string
      enum: [ SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED, NOT_SHOWED_UP ]
      description: "Status of the visit"
    VeterinarianSchedule:
      type: object
      properties:
        veterinarian_id:
          $ref: '#/components/schemas/VeterinarianId'
        start_date:
          $ref: '#/components/schemas/LocalDate'
        end_date:
          $ref: '#/components/schemas/LocalDate'
        visits:
          type: array
          items:
            $ref: '#/components/schemas/Visit'
        availabilities:
          type: array
          items:
            $ref: '#/components/schemas/Availability'
...
    VeterinarianDemand:
      type: object
      properties:
        veterinarian_id:
          $ref: '#/components/schemas/VeterinarianId'
        first_name:
          type: string
        last_name:
          type: string
        scheduled_visits:
          type: integer
          format: int64
          description: "Number of scheduled visits in the requested period"
...
paths:
...
  /veterinarians/{vet_id}/availabilities:
    post:
      tags: [ Veterinarians ]
      summary: "Create a new availability for a Veterinarian"
      security:
        - bearerAuth: [RECEPTIONIST]
      parameters:
        - $ref: '#/components/parameters/vet_id'
      requestBody:
        $ref: '#/components/requestBodies/AvailabilityInput'
      responses:
        '201':
          description: "Availability created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
        '400':
          description: "Invalid input"
        '404':
          description: "Veterinarian not found"
    get:
      tags: [ Veterinarians ]
      summary: "Get all availabilities from a Veterinarian"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/vet_id'
      responses:
        '200':
          description: "Availabilities found"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Availability'
        '404':
          description: "Availabilities not found"
    delete:
      tags: [ Veterinarians ]
      summary: "Delete all availabilities from a Veterinarian"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/vet_id'
      responses:
        '204':
          description: "Availabilities deleted"
        '404':
          description: "Veterinarian not found"
  /veterinarians/{vet_id}/schedule:
    get:
      tags: [ Veterinarians ]
      summary: "Get the schedule of a Veterinarian"
      security:
        - bearerAuth: [VETERINARIAN, RECEPTIONIST]
      parameters:
        - $ref: '#/components/parameters/vet_id'
        - $ref: '#/components/parameters/start_date'
        - $ref: '#/components/parameters/end_date'
      responses:
        '200':
          description: "Schedule found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VeterinarianSchedule'
        '400':
          description: "Invalid date range"
        '404':
          description: "Veterinarian not found"
  /veterinarians/demand:
    get:
      tags: [ Veterinarians ]
      summary: "List veterinarians ordered by demand in a given period"
      description: "Returns veterinarians ordered descending by the number of scheduled visits in the given period."
      security:
        - bearerAuth: [ CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/start_date'
        - $ref: '#/components/parameters/end_date'
      responses:
        '200':
          description: "List of veterinarians ordered by demand"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VeterinarianDemand'
        '400':
          description: "Invalid date range"
  /veterinarians/{vet_id}/availabilities/{availability_id}:
    get:
      tags: [ Veterinarians ]
      summary: "Get an availability"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/availability_id'
      responses:
        '200':
          description: "Availability found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
        '404':
          description: "Availability not found"
    put:
      tags: [ Veterinarians ]
      summary: "Update an availability"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/availability_id'
      requestBody:
        $ref: '#/components/requestBodies/AvailabilityInput'
      responses:
        '200':
          description: "Availability updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
        '400':
          description: "Invalid input"
        '404':
          description: "Availability not found"
    delete:
      tags: [ Veterinarians ]
      summary: "Delete an availability"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/availability_id'
      responses:
        '204':
          description: "Availability deleted"
        '404':
          description: "Availability not found"
  /veterinarians/{vet_id}/availabilities/{availability_id}/exceptions:
    post:
      tags: [ Veterinarians ]
      summary: "Create a new exception for an availability"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/availability_id'
      requestBody:
        $ref: '#/components/requestBodies/AvailabilityExceptionInput'
      responses:
        '201':
          description: "Exception created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityException'
        '400':
          description: "Invalid input"
        '404':
          description: "Availability not found"
    get:
      tags: [ Veterinarians ]
      summary: "Get all exceptions from an Availability"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/availability_id'
      responses:
        '200':
          description: "Exceptions found"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailabilityException'
        '404':
          description: "Exceptions not found"
    delete:
      tags: [ Veterinarians ]
      summary: "Delete all exceptions from an Availability"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/availability_id'
      responses:
        '204':
          description: "Exceptions deleted"
        '404':
          description: "Availability not found"
  /veterinarians/{vet_id}/availabilities/{availability_id}/exceptions/{exception_id}:
    get:
      tags: [ Veterinarians ]
      summary: "Get an exception"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/exception_id'
      responses:
        '200':
          description: "Exception found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityException'
        '404':
          description: "Exception not found"
    put:
      tags: [ Veterinarians ]
      summary: "Update an exception"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/exception_id'
      requestBody:
        $ref: '#/components/requestBodies/AvailabilityExceptionInput'
      responses:
        '200':
          description: "Exception updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityException'
        '400':
          description: "Invalid input"
        '404':
          description: "Exception not found"
    delete:
      tags: [ Veterinarians ]
      summary: "Delete an exception"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/exception_id'
      responses:
        '204':
          description: "Exception deleted"
        '404':
          description: "Exception not found"

  # ============================================
  # Medications
  # ============================================
...